requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
pydantic-settings==2.10.1
python-dotenv==1.0.0
openai==1.42.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
openai==1.42.0